JSON wire protocol hacking
Read JSON serialized procedure invocations from clients that
connect with a "json" service string (vs. "database" or "export").

Input document is (int) length preceeded on the wire for
framing. The procedure invocation document looks like:

{
   procedure : "procedure_name",
   handle : long,
   partition-key : (Number | String),
   parameters : []
}

partition-key is a string or number json value providing
the value of the partitioning column. The partition column
is still provided in the parameters array as usual. It
is extracted here largely as an optimization to avoid
deserializing the parameters in the ClientInterface and
as an attempt at greater clarity vs. the native API.

This is work in progress.
 * All responses are still VoltTables, including all error responses.
 * Establishing a new client connection still requires the usual
   wire-protocol dance.