Log IV2 trace messages in a separate thread.
This is the first step of making the trace app. Logging of IV2 trace messages
need to happen in a separate thread in order not to block execution.

I created a generic data structure to hold trace messages so that it doesn't
matter if the original messages are modified before they're logged.

For now, there is only one thread logging to log4j. It's easy to increase the
thread count.

Unless IV2TRACE or IV2QUEUETRACE is set to TRACE level, it won't start the
thread nor log anything.